Populate placeholders in content when specified
Placeholders in content files are not populated by default. If
placeholders are to be populated in content files, parameter named
`render` with a string value of `yes` must be specified.

This `render` parameter may be specified as a header in content files or
it may be specified as a keyword argument in `make_pages` call.